requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
pdfplumber
Pillow 
firebase-admin 
openpyxl